+++ Example/Example/Models/Virtual Currency List/VirtualCurrencyList.swift
@@ -64,7 +64,7 @@ class VirtualCurrencyList
     {
         logger.info { "Buy action for item: \(item.name)" }
 
-        let uiSettings = StorePaymentProjectSettings.UISettings(theme: AppConfig.paystationUITheme,
+        let uiSettings = StorePaymentProjectSettings.UISettings(theme: AppConfig.paystationUIThemeId,
                                                                 size: AppConfig.paystationUISize)
         let redirectPolicy =
             StorePaymentProjectSettings.RedirectPolicy(redirectConditions: .any,