+++ StoreKit/Models/StoreVirtualItem.swift
@@ -51,10 +51,10 @@ public struct StoreVirtualItem
 
     /// Type of virtual item.
     public let virtualItemType: String
-    
+
     /// Promotion settings in Store.
     public let promotions: [StoreItemPromotion]
-    
-    /// Limits.
+
+    /// Limit to the number of times one user can buy an item.
     public let limits: StoreItemLimits?
 }